SignalR update
Reconfigured the logic for signalR connection. Before the connection was starting when a chat was opened, but now connection is opened right after the user would log in / or go to other page, with all the friends he has. This helps the messages to go smooth and enables for other communication between the users, ex. Sending notifications, likes, comments etc