Image Like / Image carousel update
Implemented likes for gallery images.
Notification is sent live to the other user to notify him that a friend ( it shows the full name of the friend ) liked/disliked an image of his ( shows which image ).
Changed the implementation of the carousel. Instead of generating the carousel, everything is printed out in the cshtml file, and when clicking an image, it toggles the carousel from class=hide to class=full, and it scrolls to the clicked image.
It must be like this, since before when you open an image it gives the current image likes and comments, and that aplies for the rest of the images in carousel.